feat: header, footer styles 분리
