+++ src/components/Footer/Footer.tsx
@@ -1,8 +1,10 @@
-import {View, StyleSheet, Text, Image} from 'react-native';
+import {View, Text, Image} from 'react-native';
 import React from 'react';
 
 const facebookImage = require('../../../assets/mobile_semtle_common_asset/footer/footer_facebook_icon.png');
 const instaImage = require('../../../assets/mobile_semtle_common_asset/footer/footer_insta_icon.png');
+
+import styles from './styles';
 const Footer = () => {
   return (
     <View style={styles.footer}>
@@ -18,32 +20,5 @@ const Footer = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  footer: {
-    position: 'absolute',
-    bottom: 0,
-    width: '100%',
-    height: 132,
-    backgroundColor: '#f5f5f5',
-  },
-  footerText: {
-    textAlign: 'center',
-    fontSize: 9,
-    lineHeight: 16,
-    fontWeight: '400',
-  },
-  footerIcon: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-  },
-  footerImage: {
-    width: 12,
-    height: 12,
-  },
-  iconSeparator: {
-    width: 10, // 간격을 조정하기 위한 너비
-  },
-});
 
 export default Footer;
